Refactorization API Consultation

+++ src/main/java/com/hegaro/medicalapp/repository/ConsultationRepository.java
@@ -3,5 +3,5 @@ package com.hegaro.medicalapp.repository;
 import com.hegaro.medicalapp.model.Consultation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ConsultationRepository extends JpaRepository<Consultation, Integer> {
+public interface ConsultationRepository extends JpaRepository<Consultation, Long> {
 }
